+++ Sources/Yams/Emitter.swift
@@ -41,7 +41,8 @@ public func dump<Objects>(
     version: (major: Int, minor: Int)? = nil,
     sortKeys: Bool = false,
     sequenceStyle: Node.Sequence.Style = .any,
-    mappingStyle: Node.Mapping.Style = .any) throws -> String
+    mappingStyle: Node.Mapping.Style = .any,
+    newLineScalarStyle: Node.Scalar.Style = .any) throws -> String
     where Objects: Sequence {
     func representable(from object: Any) throws -> NodeRepresentable {
         if let representable = object as? NodeRepresentable {
@@ -62,7 +63,8 @@ public func dump<Objects>(
         version: version,
         sortKeys: sortKeys,
         sequenceStyle: sequenceStyle,
-        mappingStyle: mappingStyle
+        mappingStyle: mappingStyle,
+        newLineScalarStyle: newLineScalarStyle
     )
 }
 
@@ -96,7 +98,9 @@ public func dump(
     version: (major: Int, minor: Int)? = nil,
     sortKeys: Bool = false,
     sequenceStyle: Node.Sequence.Style = .any,
-    mappingStyle: Node.Mapping.Style = .any) throws -> String {
+    mappingStyle: Node.Mapping.Style = .any,
+    newLineScalarStyle: Node.Scalar.Style = .any,
+    redundancyAliasingStrategy: RedundancyAliasingStrategy? = nil) throws -> String {
     return try serialize(
         node: object.represented(),
         canonical: canonical,
@@ -109,7 +113,9 @@ public func dump(
         version: version,
         sortKeys: sortKeys,
         sequenceStyle: sequenceStyle,
-        mappingStyle: mappingStyle
+        mappingStyle: mappingStyle,
+        newLineScalarStyle: newLineScalarStyle,
+        redundancyAliasingStrategy: redundancyAliasingStrategy
     )
 }
 
@@ -143,7 +149,9 @@ public func serialize<Nodes>(
     version: (major: Int, minor: Int)? = nil,
     sortKeys: Bool = false,
     sequenceStyle: Node.Sequence.Style = .any,
-    mappingStyle: Node.Mapping.Style = .any) throws -> String
+    mappingStyle: Node.Mapping.Style = .any,
+    newLineScalarStyle: Node.Scalar.Style = .any,
+    redundancyAliasingStrategy: RedundancyAliasingStrategy? = nil) throws -> String
     where Nodes: Sequence, Nodes.Iterator.Element == Node {
     let emitter = Emitter(
         canonical: canonical,
@@ -156,7 +164,9 @@ public func serialize<Nodes>(
         version: version,
         sortKeys: sortKeys,
         sequenceStyle: sequenceStyle,
-        mappingStyle: mappingStyle
+        mappingStyle: mappingStyle,
+        newLineScalarStyle: newLineScalarStyle,
+        redundancyAliasingStrategy: redundancyAliasingStrategy
     )
     try emitter.open()
     try nodes.forEach(emitter.serialize)
@@ -194,7 +204,9 @@ public func serialize(
     version: (major: Int, minor: Int)? = nil,
     sortKeys: Bool = false,
     sequenceStyle: Node.Sequence.Style = .any,
-    mappingStyle: Node.Mapping.Style = .any) throws -> String {
+    mappingStyle: Node.Mapping.Style = .any,
+    newLineScalarStyle: Node.Scalar.Style = .any,
+    redundancyAliasingStrategy: RedundancyAliasingStrategy? = nil) throws -> String {
     return try serialize(
         nodes: [node],
         canonical: canonical,
@@ -207,7 +219,9 @@ public func serialize(
         version: version,
         sortKeys: sortKeys,
         sequenceStyle: sequenceStyle,
-        mappingStyle: mappingStyle
+        mappingStyle: mappingStyle,
+        newLineScalarStyle: newLineScalarStyle,
+        redundancyAliasingStrategy: redundancyAliasingStrategy
     )
 }
 
@@ -238,10 +252,10 @@ public final class Emitter {
         public var allowUnicode: Bool = false
         /// Set the preferred line break.
         public var lineBreak: LineBreak = .ln
-
-        // internal since we don't know if these should be exposed.
-        var explicitStart: Bool = false
-        var explicitEnd: Bool = false
+        /// Set to emit an explicit document start marker.
+        public var explicitStart: Bool = false
+        /// Set to emit an explicit document end marker.
+        public var explicitEnd: Bool = false
 
         /// The `%YAML` directive value or nil.
         public var version: (major: Int, minor: Int)?
@@ -254,9 +268,52 @@ public final class Emitter {
 
         /// Set the style for mappings (dictionaries)
         public var mappingStyle: Node.Mapping.Style = .any
+
+        /// Set the style for scalars that include newlines
+        public var newLineScalarStyle: Node.Scalar.Style = .any
         
         /// Redundancy aliasing strategy to use when encoding. Defaults to nil
         public var redundancyAliasingStrategy: RedundancyAliasingStrategy?
+        
+        /// Create `Emitter.Options` with the specified values.
+        ///
+        /// - parameter canonical:     Set if the output should be in the "canonical" format described in the YAML
+        ///                            specification.
+        /// - parameter indent:        Set the indentation value.
+        /// - parameter width:         Set the preferred line width. -1 means unlimited.
+        /// - parameter allowUnicode:  Set if unescaped non-ASCII characters are allowed.
+        /// - parameter lineBreak:     Set the preferred line break.
+        /// - parameter explicitStart: Explicit document start `---`.
+        /// - parameter explicitEnd:   Explicit document end `...`.
+        /// - parameter version:       The `%YAML` directive value or nil.
+        /// - parameter sortKeys:      Set if emitter should sort keys in lexicographic order.
+        /// - parameter sequenceStyle: Set the style for sequences (arrays / lists)
+        /// - parameter mappingStyle:  Set the style for mappings (dictionaries)
+        /// - parameter newLineScalarStyle: Set the style for newline-containing scalars
+        /// - parameter redundancyAliasingStrategy: Set the strategy for identifying redundant structures and automatically aliasing them
+        public init(canonical: Bool = false, indent: Int = 0, width: Int = 0, allowUnicode: Bool = false,
+                    lineBreak: Emitter.LineBreak = .ln,
+                    explicitStart: Bool = false,
+                    explicitEnd: Bool = false,
+                    version: (major: Int, minor: Int)? = nil,
+                    sortKeys: Bool = false, sequenceStyle: Node.Sequence.Style = .any,
+                    mappingStyle: Node.Mapping.Style = .any,
+                    newLineScalarStyle: Node.Scalar.Style = .any,
+                    redundancyAliasingStrategy: RedundancyAliasingStrategy? = nil) {
+            self.canonical = canonical
+            self.indent = indent
+            self.width = width
+            self.allowUnicode = allowUnicode
+            self.lineBreak = lineBreak
+            self.explicitStart = explicitStart
+            self.explicitEnd = explicitEnd
+            self.version = version
+            self.sortKeys = sortKeys
+            self.sequenceStyle = sequenceStyle
+            self.mappingStyle = mappingStyle
+            self.newLineScalarStyle = newLineScalarStyle
+            self.redundancyAliasingStrategy = redundancyAliasingStrategy
+        }
     }
 
     /// Configuration options to use when emitting YAML.
@@ -280,6 +337,8 @@ public final class Emitter {
     /// - parameter sortKeys:      Set if emitter should sort keys in lexicographic order.
     /// - parameter sequenceStyle: Set the style for sequences (arrays / lists)
     /// - parameter mappingStyle:  Set the style for mappings (dictionaries)
+    /// - parameter newLineScalarStyle: Set the style for newline-containing scalars
+    /// - parameter redundancyAliasingStrategy: Set the strategy for identifying redundant structures and automatically aliasing them
     public init(canonical: Bool = false,
                 indent: Int = 0,
                 width: Int = 0,
@@ -290,7 +349,9 @@ public final class Emitter {
                 version: (major: Int, minor: Int)? = nil,
                 sortKeys: Bool = false,
                 sequenceStyle: Node.Sequence.Style = .any,
-                mappingStyle: Node.Mapping.Style = .any) {
+                mappingStyle: Node.Mapping.Style = .any,
+                newLineScalarStyle: Node.Scalar.Style = .any,
+                redundancyAliasingStrategy: RedundancyAliasingStrategy? = nil) {
         options = Options(canonical: canonical,
                           indent: indent,
                           width: width,
@@ -301,7 +362,9 @@ public final class Emitter {
                           version: version,
                           sortKeys: sortKeys,
                           sequenceStyle: sequenceStyle,
-                          mappingStyle: mappingStyle)
+                          mappingStyle: mappingStyle,
+                          newLineScalarStyle: newLineScalarStyle,
+                          redundancyAliasingStrategy: redundancyAliasingStrategy)
         // configure emitter
         yaml_emitter_initialize(&emitter)
         yaml_emitter_set_output(&self.emitter, { pointer, buffer, size in
@@ -403,42 +466,10 @@ public final class Emitter {
     }
 }
 
-// MARK: - Options Initializer
+//// MARK: - Options Initializer
 
 extension Emitter.Options {
-    /// Create `Emitter.Options` with the specified values.
-    ///
-    /// - parameter canonical:     Set if the output should be in the "canonical" format described in the YAML
-    ///                            specification.
-    /// - parameter indent:        Set the indentation value.
-    /// - parameter width:         Set the preferred line width. -1 means unlimited.
-    /// - parameter allowUnicode:  Set if unescaped non-ASCII characters are allowed.
-    /// - parameter lineBreak:     Set the preferred line break.
-    /// - parameter explicitStart: Explicit document start `---`.
-    /// - parameter explicitEnd:   Explicit document end `...`.
-    /// - parameter version:       The `%YAML` directive value or nil.
-    /// - parameter sortKeys:      Set if emitter should sort keys in lexicographic order.
-    /// - parameter sequenceStyle: Set the style for sequences (arrays / lists)
-    /// - parameter mappingStyle:  Set the style for mappings (dictionaries)
-    public init(canonical: Bool = false, indent: Int = 0, width: Int = 0, allowUnicode: Bool = false,
-                lineBreak: Emitter.LineBreak = .ln,
-                explicitStart: Bool = false,
-                explicitEnd: Bool = false,
-                version: (major: Int, minor: Int)? = nil,
-                sortKeys: Bool = false, sequenceStyle: Node.Sequence.Style = .any,
-                mappingStyle: Node.Mapping.Style = .any) {
-        self.canonical = canonical
-        self.indent = indent
-        self.width = width
-        self.allowUnicode = allowUnicode
-        self.lineBreak = lineBreak
-        self.explicitStart = explicitStart
-        self.explicitEnd = explicitEnd
-        self.version = version
-        self.sortKeys = sortKeys
-        self.sequenceStyle = sequenceStyle
-        self.mappingStyle = mappingStyle
-    }
+    
 }
 
 // MARK: Implementation Details