+++ Tests/YamsTests/ConstructorTests.swift
@@ -515,29 +515,17 @@ class ConstructorTests: XCTestCase { // swiftlint:disable:this type_body_length
         XCTAssertEqual(nodes[1]["link with"]?[1]?["="], "library2.dll")
         XCTAssertEqual(nodes[1]["link with"]?[1]?["version"], "2.3")
     }
+
+    func testStrictParsingSexagesimal() throws {
+        XCTAssertEqual(Int.construct(from: .init("12:34:56")), 45296)
+        XCTAssertEqual(Double.construct(from: .init("12:34:56.789")), 45296.789)
+        XCTAssertNil(Int.construct(from: .init("http://example.com")))
+        XCTAssertNil(Double.construct(from: .init("http://example.com")))
+    }
 }
 
 extension ConstructorTests {
     static var allTests: [(String, (ConstructorTests) -> () throws -> Void)] {
-#if os(Windows) && swift(<5.6)
-        return [
-            ("testBinary", testBinary),
-            ("testBool", testBool),
-            // ("testFloat", testFloat),
-            // ("testInt", testInt),
-            ("testMap", testMap),
-            // ("testMerge", testMerge),
-            ("testNull", testNull),
-            // ("testOmap", testOmap),
-            ("testPairs", testPairs),
-            ("testQuotationMark", testQuotationMark),
-            ("testSet", testSet),
-            ("testSeq", testSeq),
-            ("testTimestamp", testTimestamp),
-            ("testTimestampWithNanosecond", testTimestampWithNanosecond)
-            // ("testValue", testValue)
-        ]
-#else
         return [
             ("testBinary", testBinary),
             ("testBool", testBool),
@@ -555,6 +543,5 @@ extension ConstructorTests {
             ("testTimestampWithNanosecond", testTimestampWithNanosecond),
             ("testValue", testValue)
         ]
-#endif
     }
 } // swiftlint:disable:this file_length